+++ NEWS.md
@@ -3,4 +3,4 @@ centerline 0.0.0.9000 (2024-03-01)
 
 ### NEW FEATURES
 
-  * Hello World!
+  * Core functionality is added through the `cent_skeleton` function